requirements.txt
pulp>=2.8
pandas>=2.2
pyarrow>=15
numpy>=1.26
scipy>=1.12
plotly>=5.20
pydeck>=0.9
streamlit>=1.32
python-pptx>=0.6.23
matplotlib>=3.8
jupyter>=1.0
tqdm>=4.66
